Changes in the project approach
Triple class has been changed to Coordinate class. Now, there are doubles and not integers
Matrix class added to make the rotations of the points with the "multiply" method
NeuralNetworkConsole added: is the class that will launch the application with the "run" method

+++ src/com/company/SvgController.java
@@ -8,10 +8,10 @@ public class SvgController {
     int scaleY=2;
     int scaleZ=5;
     String color;
-    int length=100;
-    int lenghtAux=0;
-    int auxY;
-    int auxX;
+    double length=100;
+    double lenghtAux=0;
+    double auxY;
+    double auxX;
     int i;
     String svgString="<html>\n" +
             "\t"+"<head>\n" +
